Create College Admission Form
It is a simple project done with Python code. Here I have created an College Admission Form by using Tk inter package  in Python where i had written 120 lines of code to do it so. Further I had integerated MySql to save my entered data in the Form and created a seperate database and Tables to save my details. This is a sample example of how the webpages work and how they can save the user data in their databases.  We can retrive the Data anytime using MySql. This is only a sample. Futher we can create a webpage interface by using adobe photoshop etc... Thanks for seeing my post.